move DBL sign downloader class to its own file
